Fixed a tiny issue
I'm heading to your place to program, still, had to do it to em

+++ src/ConsoleUI/ConsoleUI.cpp
@@ -6,7 +6,7 @@ ConsoleUI::ConsoleUI(vector<string> places){
     cout<<flush;
     system("cls");
     outputList(places);
-    cursorPosition = {X:0,Y:places.size()};
+    cursorPosition = {X:0,Y:(short)places.size()};
     allPlaces = places;
 }
